Removed const_marker_extra_impls feature, now those impls are always enabled
