+++ src/const_marker.rs
@@ -50,13 +50,6 @@ mod const_marker_eq_traits;
 pub use const_marker_eq_traits::*;
 
 
-#[cfg(feature = "const_marker_extra_impls")]
-mod const_marker_additional_impls;
-
-#[cfg(feature = "const_marker_extra_impls")]
-pub(crate) use const_marker_additional_impls::*;
-
-
 #[cfg(feature = "serde")]
 mod const_marker_serde_impls;
 
@@ -236,8 +229,26 @@ macro_rules! __declare_const_param_type {
             }
         }
 
-        #[cfg(feature = "const_marker_extra_impls")]
-        crate::const_marker::__declare_const_marker_additional_impls!{$struct($prim)}
+        impl<const L: $prim, const R: $prim> core::cmp::PartialOrd<$struct<R>> for $struct<L> {
+            fn partial_cmp(&self, _: &$struct<R>) -> Option<core::cmp::Ordering> {
+                Some(core::cmp::Ord::cmp(&L, &R))
+            }
+        }
+
+        impl<const VAL: $prim> core::cmp::Ord for $struct<VAL> {
+            fn cmp(&self, _: &Self) -> core::cmp::Ordering {
+                core::cmp::Ordering::Equal
+            }
+        }
+
+        impl<const VAL: $prim> core::hash::Hash for $struct<VAL> {
+            fn hash<H>(&self, state: &mut H)
+            where
+                H: core::hash::Hasher,
+            {
+                core::hash::Hash::hash(&VAL, state)
+            }
+        }
 
         #[cfg(feature = "serde")]
         crate::const_marker::__declare_const_marker_serde_impls!{$struct($prim)}
